replace with 0.11.3

+++ medusa/purge.py
@@ -21,13 +21,14 @@ import traceback
 
 from datetime import datetime, timedelta
 
+import medusa.utils
 from medusa.index import clean_backup_from_index
 from medusa.monitoring import Monitoring
 from medusa.storage import Storage, format_bytes_str
 
 
 def main(config, max_backup_age=0, max_backup_count=0):
-    backups_to_purge = list()
+    backups_to_purge = set()
     monitoring = Monitoring(config=config.monitoring)
 
     try:
@@ -38,11 +39,11 @@ def main(config, max_backup_age=0, max_backup_count=0):
         backup_index = storage.list_backup_index_blobs()
         backups = list(storage.list_node_backups(fqdn=config.storage.fqdn, backup_index_blobs=backup_index))
         # list all backups to purge based on date conditions
-        backups_to_purge += backups_to_purge_by_age(backups, max_backup_age)
+        backups_to_purge |= set(backups_to_purge_by_age(backups, max_backup_age))
         # list all backups to purge based on count conditions
-        backups_to_purge += backups_to_purge_by_count(backups, max_backup_count)
+        backups_to_purge |= set(backups_to_purge_by_count(backups, max_backup_count))
         # purge all candidate backups
-        purge_backups(storage, backups_to_purge)
+        purge_backups(storage, backups_to_purge, config.storage.backup_grace_period_in_days)
 
         logging.debug('Emitting metrics')
         tags = ['medusa-node-backup', 'purge-error', 'PURGE-ERROR']
@@ -82,9 +83,9 @@ def backups_to_purge_by_count(backups, max_backup_count):
 #     return list(filter(lambda backup: backup.name = backup_name, backups)) or list()
 
 
-def purge_backups(storage, backups):
+def purge_backups(storage, backups, backup_grace_period_in_days):
     """
-    Core function to purge a list of node_backups
+    Core function to purge a set of node_backups
     Used for node purge and backup delete (using a specific backup_name)
     """
     logging.info("{} backups are candidate to be purged".format(len(backups)))
@@ -99,7 +100,9 @@ def purge_backups(storage, backups):
         fqdns.add(backup.fqdn)
 
     for fqdn in fqdns:
-        (cleaned_objects_count, cleaned_objects_size) = cleanup_obsolete_files(storage, fqdn)
+        (cleaned_objects_count, cleaned_objects_size) = cleanup_obsolete_files(storage,
+                                                                               fqdn,
+                                                                               backup_grace_period_in_days)
         nb_objects_purged += cleaned_objects_count
         total_purged_size += cleaned_objects_size
 
@@ -127,33 +130,46 @@ def purge_backup(storage, backup):
     return (purged_objects, purged_size)
 
 
-def cleanup_obsolete_files(storage, fqdn):
+def cleanup_obsolete_files(storage, fqdn, backup_grace_period_in_days):
     logging.info("Cleaning up orphaned files for {}...".format(fqdn))
     nb_objects_purged = 0
     total_purged_size = 0
 
     backups = storage.list_node_backups(fqdn=fqdn)
     paths_in_manifest = get_file_paths_from_manifests_for_differential_backups(backups)
-    paths_in_storage = get_file_paths_from_storage(storage, fqdn)
+    paths_in_storage = get_file_paths_from_storage(storage, fqdn, backup_grace_period_in_days)
 
     for path in paths_in_storage - paths_in_manifest:
         logging.debug("  - [{}] exists in storage, but not in manifest".format(path))
         obj = storage.storage_driver.get_blob(path)
-        nb_objects_purged += 1
-        total_purged_size += int(obj.size)
-        storage.storage_driver.delete_object(obj)
+        if obj is not None:
+            nb_objects_purged += 1
+            total_purged_size += int(obj.size)
+            storage.storage_driver.delete_object(obj)
 
     return nb_objects_purged, total_purged_size
 
 
-def get_file_paths_from_storage(storage, fqdn):
-    data_directory = "{}/data".format(fqdn)
+def get_file_paths_from_storage(storage, fqdn, backup_grace_period_in_days):
+    data_directory = "{}{}/data".format(storage.prefix_path, fqdn)
     data_files = {
         blob.name: blob
-        for blob in storage.storage_driver.list_objects(str(data_directory))
+        for blob in filter_files_within_gc_grace(storage.storage_driver,
+                                                 storage.storage_driver.list_objects(str(data_directory)),
+                                                 backup_grace_period_in_days)
     }
 
     return set(data_files.keys())
+
+
+def filter_files_within_gc_grace(storage_driver, blobs, backup_grace_period_in_days):
+    return list(
+        filter(lambda blob:
+               is_older_than_gc_grace(storage_driver.get_object_datetime(blob), backup_grace_period_in_days), blobs))
+
+
+def is_older_than_gc_grace(blob_datetime, gc_grace) -> bool:
+    return datetime.timestamp(blob_datetime) <= datetime.timestamp(datetime.now()) - (int(gc_grace) * 86400)
 
 
 def get_file_paths_from_manifests_for_differential_backups(backups):
@@ -193,14 +209,16 @@ def delete_backup(config, backup_name, all_nodes):
             backups_to_purge = [nb for nb in backups_to_purge if storage.config.fqdn in nb.fqdn]
 
         logging.info('Deleting Backup {}...'.format(backup_name))
-        purge_backups(storage, backups_to_purge)
+        purge_backups(storage, backups_to_purge, config.storage.backup_grace_period_in_days)
 
         logging.debug('Emitting metrics')
         tags = ['medusa-node-backup', 'delete-error', 'DELETE-ERROR']
         monitoring.send(tags, 0)
     except Exception as e:
-        traceback.print_exc()
         tags = ['medusa-node-backup', 'delete-error', 'DELETE-ERROR']
         monitoring.send(tags, 1)
-        logging.error('This error happened during the delete of backup "{}": {}'.format(backup_name, str(e)))
-        sys.exit(1)
+        medusa.utils.handle_exception(
+            e,
+            'This error happened during the delete of backup "{}": {}'.format(backup_name, str(e)),
+            config
+        )
